requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
apscheduler
Pillow
python-multipart
